bldprogs/VBoxEditElf: Experiment with creating stub linker libraries, bugref:10874
The tool parses a given shared library (.so), extracts the information relevant for linking
and writes the result into a new file which acts as a stub. This way .so files can get much smaller
and be distributed in an SDK style tools package without requiring a lot of disk space.



git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108509 cfe28804-0f27-0410-a406-dd0f0b0b656f